+++ src2/module/files/head.py
@@ -1,107 +1,4 @@
-from os import path
-
-from src2.services.images import format_sizes
-from src2.helpers import FilesValidator, KeysValidator
-
-
-class Images:
-    icons_config: dict
-    config: dict
-
-    def _requirements(self, key):
-        if key not in self.config:
-            return False
-        KeysValidator(key, value=self.config[key]).key_is_not_void()
-        file_path = path.join(self.config.get('main_path', ''),
-                              self.config[key])
-        FilesValidator(key, file_path).path_is_file()
-        return True
-
-    def favicon_png(self):
-        head = []
-        if not self._requirements('favicon_png'): return head
-        for icon_brand_config in self.icons_config.get('favicon_png', []):
-            sizes = format_sizes(icon_brand_config)
-            for size in sizes:
-                filename = (f"{icon_brand_config['filename']}-"
-                            f"{size[0]}x{size[1]}.png")
-                element = self._icons_head_creator(filename, icon_brand_config,
-                                                   size)
-                head.append(element)
-        return head
-
-    def favicon_ico(self):
-        return [( "<link rel='shortcut icon' "
-                 f"href='/favicon.ico' type='image/x-icon' />")]
-
-    def favicon_svg(self):
-        color = self.config.get('background_color', '')
-        filename = self.config.get('favicon_svg', '')
-        return [f"<link rel='mask-icon' href='{filename}' color='{color}' />"]
-
-    def preview_png(self):
-        # og:image (http), og:image:secure_url (https) and twitter:image
-        static_url = self.config.get('static_url', '')
-        image = f"{static_url}preview.png"
-        return [
-            f"<meta property='og:image' content='{image}' />",
-            f"<meta property='og:image:secure_url' content='{image}' />",
-            f"<meta name='twitter:image' content='{image}' />",
-        ]
-
-    def _icons_head_creator(self, filename, icon_brand_config, size):
-        min_size = size[1]
-        if 'media' in icon_brand_config:
-            size[1] = size[0]
-        name_ref = icon_brand_config.get('name_ref', '')
-        static_url = self.config['static_url']
-        file_type = (
-            f"type='{icon_brand_config['file_type']}' "
-            if 'file_type' in icon_brand_config else ''
-        )
-        sizes = (
-            f"sizes='{size[0]}x{size[1]}' "
-            if 'verbosity' in icon_brand_config else ''
-        )
-        tagname, attribute, ref = (
-            ('meta', 'name', 'content')
-            if 'metatag' in icon_brand_config else
-            ('link', 'rel', 'href')
-        )
-        title = (
-            f"title='{icon_brand_config['title']}' "
-            if 'title' in icon_brand_config else ''
-        )
-        media = (
-            f"media=(min-device-width: {min_size}px) and "
-            f"(min-device-height: {min_size}px) "
-            if 'media' in icon_brand_config else ''
-        )
-        if 'content' in icon_brand_config:
-            static_url = ''
-            filename = f"content='{icon_brand_config['content']}'"
-
-        # Keep using format function for better reference to each element of
-        # the formated string
-        # A: <link rel="shortcut icon" type="image/png" sizes="16x16"
-        #    href="/static/favicon-16x16.png" />
-        # B: <link href="apple-touch-startup-image-320x320.png"
-        #    media="screen and (min-device-width: 320px) and ... ..." />
-        # C: <link rel="fluid-icon" href="/static/fluidicon-512x512.png"
-        #    title="Microsoft" />
-        element = "<{} {}='{}' {}{}{}='{}{}' {}{}/>".format(
-            tagname,  # A: link
-            attribute,  # A: rel
-            name_ref,  # A: shortcut icon
-            file_type,  # A: type="image/png"
-            sizes,  # A: sizes="16x16"
-            ref,  # A: href
-            static_url,  # A: /static/
-            filename,  # A: favicon-16x16.png
-            media,  # B: media="screen and (min-device-width: 320px) and ...
-            title,  # C: title="Microsoft"
-        )
-        return element
+from .images import Images
 
 class Head(Images):
     config = {}
@@ -286,67 +183,3 @@ class Head(Images):
         ]
         return head
 
-class Index(Head):
-    
-    def full_index(self):
-        head = self.full_head()
-        html = self.structure(head)
-        return html
-    
-    def structure(self, head):
-        indent = "    "  # 4 spaces
-        formated_head = ''.join([
-            f"{indent*2}{tag}\n"
-            for conjunt in head
-            for tag in conjunt
-        ])
-        return ( "<html>\n"
-                f"{indent}<head>\n"
-                f"{formated_head}"  # Already have newline
-                f"{indent}</head>\n"
-                 "</html>")
-
-
-class ComplementaryFiles:
-    icons_config = {}
-    config = {}
-
-    def browserconfig(self):
-        """browserconfig.xml"""
-        if not 'browserconfig' in self.icons_config:
-            return ''
-
-        config = self.icons_config['browserconfig']
-        icon = config.get('filename', '')
-        square_sizes = config.get('square_sizes', [])
-        non_square_sizes = config.get('non_square_sizes', [])
-        static_url = self.config.get('static_url', '')
-
-        browserconfig = ("<?xml version='1.0' encoding='utf-8'?><browserconfig>"
-                         "<msapplication><tile>")
-        browserconfig += ''.join([
-            "<square{0}x{0}logo src='{1}{2}-{0}x{0}.png' />".format(
-                size,
-                static_url,
-                icon
-            )
-            for size in square_sizes
-        ])
-        browserconfig += ''.join([
-            "<wide{0}x{1}logo src='{2}{3}-{0}x{1}.png' />".format(
-                size[0],
-                size[1],
-                static_url,
-                icon
-            )
-            for size in non_square_sizes
-        ])
-        color = self.config.get('background_color', '')
-        browserconfig += (f"<TileColor>{color}</TileColor></tile></msapplication>"
-                           "</browserconfig>")
-        browserconfig = browserconfig.replace('\'', '"')
-        return browserconfig
-
-
-class Files(Index, ComplementaryFiles):
-    pass
